Move detector methods into standalone files with partial class
